Added and integrated implicit preloading scan on model load.
When loading a model from a text file, the assumption was made
that doing an initial scan to get a linecount, in addition to
the normal parsing, would give a slight performance gain --
this seems to be the case (+8% speed for a few quick loading
tests). The reason for doing so is to reduce the resizing
of the hashmap to be filled (Model.Data).